[#39] SplashViewModel 분리 및 TokenConfig 상수화
- SplashViewModel을 별도 파일로 분리
- TokenConfig enum 추가하여 expirationBuffer 상수화
